board: marvell: Fix disable_ooo and enable_wfe smc call definition
Now these env varaiables take mask, change the parameter to 'uint64_t'
from 'unsigned int' to be able to take 36 cores coremask.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/36176
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ board/Marvell/octeontx2/smc.c
@@ -36,7 +36,7 @@ ssize_t smc_disable_rvu_lfs(unsigned int node)
 	return regs.regs[0];
 }
 
-ssize_t smc_configure_ooo(unsigned int val)
+ssize_t smc_configure_ooo(u64 val)
 {
 	struct pt_regs regs;
 
@@ -47,7 +47,7 @@ ssize_t smc_configure_ooo(unsigned int val)
 	return regs.regs[0];
 }
 
-ssize_t smc_configure_wfe(unsigned int val)
+ssize_t smc_configure_wfe(u64 val)
 {
 	struct pt_regs regs;
